app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
python-multipart>=0.0.18
pillow>=11.0.0
matplotlib>=3.9.0
